requirements.txt
Flask==3.0.0
SimpleITK==2.3.1
numpy==1.26.3
Pillow==10.2.0
pyinstaller==6.3.0
